+++ Assets/PreFabs/Platforms/Green platform.prefab
@@ -13,7 +13,7 @@ GameObject:
   - component: {fileID: 5232127324475656237}
   - component: {fileID: 6871883581717762153}
   - component: {fileID: 997492008393400127}
-  - component: {fileID: 6593599351807107600}
+  - component: {fileID: 4651214979464683065}
   m_Layer: 0
   m_Name: Green platform
   m_TagString: Green
@@ -116,7 +116,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.009809123, y: 0.008954233, z: 0.009779397}
   m_Center: {x: 0.000044062705, y: 4.656613e-10, z: -0.000000014202669}
---- !u!114 &6593599351807107600
+--- !u!114 &4651214979464683065
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -129,4 +129,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   switchState: 1
-  shouldBlink: 0
