+++ Assets/Scripts/Platform.cs
@@ -1,11 +1,18 @@
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class Platform : MonoBehaviour
 {
-    public bool isBlinking = false;
+
+    public bool switchState = true;
+    private bool _state = true;
+    // This is the time we're given for each beat
+    private float _delayTime = 3.609f;
+    // The time we should wait before platform starts blinking
+    private float _blinkDelay = 2.255f;
+
+    
     private MeshRenderer _renderer;
     private Color _startColor;
     private Color _oldColor;
@@ -16,20 +23,65 @@ public class Platform : MonoBehaviour
     
     void Start()
     {
+        
         _renderer = GetComponent<MeshRenderer>();
         _material = _renderer.materials[1];
         _startColor = _material.color;
-        _oldColor = new Color(_startColor.r, _startColor.g, _startColor.b, 1f);
-        _startColorTransp = new Color(_startColor.r, _startColor.g, _startColor.b, 0.3f);
         _collider = GetComponent<BoxCollider>();
         if (_collider == null)
         {
             _collider = GetComponent<MeshCollider>();
         }
+        
+        if (gameObject.CompareTag("Green"))
+        {
+            _state = false;
+            Disappear();
+        }
+        
+    }
+    
+    void Update()
+    {
+        if (switchState)
+        {
+            switchState = false;
+            var coroutine = SwitchState();
+            StartCoroutine(coroutine);
+        }
+        
+    }
+
+    private IEnumerator SwitchState()
+    {
+        var oldState = _state;
+        yield return new WaitForSeconds(_blinkDelay);
+        
+        // Start blink
+        if (_state)
+        {
+            for (var i = 0; i < 3; i++)
+            {
+                _material.color = _startColor * 1.5f;
+                yield return new WaitForSeconds(0.1f);
+                _material.color = _startColor;
+                yield return new WaitForSeconds(_blinkTime - 0.1f);
+            }
+            Disappear();
+        }
+        else
+        {
+            yield return new WaitForSeconds(_delayTime - _blinkDelay);
+            Reappear();
+        }
+        
+
+        _state = !oldState;
+        
+        switchState = true;
 
     }
     
-
     public void Disappear()
     {
         Color newColor = _startColor;
@@ -42,20 +94,5 @@ public class Platform : MonoBehaviour
     {
         _material.color = _startColor;
         _collider.enabled = true;
-    }
-
-    public IEnumerator Blink()
-    {
-        if (_renderer != null)
-        {
-            for (var i = 0; i < 3; i++)
-            {
-                _material.color = _startColor * 1.5f;
-                yield return new WaitForSecondsRealtime(0.1f);
-                _material.color = _startColor;
-                yield return new WaitForSecondsRealtime(_blinkTime - 0.1f);
-            }
-            isBlinking = false;
-        }
     }
 }
