--- Assets/Scripts/PlatformManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f768b46bbc9b34c9ab8711c2d3e4dca4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
