--- Assets/Scripts/PlatformTwo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d34a257b7cb50472bacc23fd981eb24c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
